pyproject.toml
[project]
name = "mtf"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
]

[tool.poetry]
packages = [{include = "mtf", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
